+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Automatically fetch user settings from previous versions of RDMP when installing the latest version
 - Allow new columns to be added to archive extractions
 - Simplify use of Extraction Progress
+- Update CHI validation to allow mod10 CHIs
 - Hardened Confluence export functionality
 
 ## [9.1.1] - 2025-12-02